+++ e2e-tests/__sandbox__/test-app-01/nested/create-nested-sandbox.js
@@ -1,4 +1,4 @@
-const Sandbox = require('../../../../src/temp-sandbox');
+const Sandbox = require('../../../../src/temp-sandbox').TempSandbox;
 
 const sandbox = new Sandbox();