+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+-   (breaking): replace default export with named export `TempSandbox`
+
 ## [1.0.18] - 2019-04-22
 
 ### Changed